a little more "radical" in that it removes the directives for create/edit in favor of states and corresponding stateful urls.
This approach has the advantage for building a proper "RESTlike" url structure. As a user, I can bookmark and generally get hints of where I am at in the application based on the state (as represented by the url).

Disadvantage of this is that you lose the nice inline editing, but you also don't have to track that state anymore, it is handled by our state machine. With some additional study, you could still get the inline editing. I think it would involve managing the state of the "current bookmark" which the app currently does not. Current bookmark would be updated when editing, and manage the process of editing the bookmark and informing interested actors about its state.

Will probably explore this a bit more...